cleanups, moving things around
